- unit tests optimization and harmonization - pmd cq bump-up - Jenkins Build Details result UI/UX improvement - Endpoint Status result UI/UX improvement

+++ src/main/java/com/javanix/bot/jenkinsBot/cli/jenkins/JenkinsBuildDetails.java
@@ -8,7 +8,7 @@ import java.util.Set;
 @Data
 @Builder
 public class JenkinsBuildDetails {
-	private Long runTestsCount;
+	private Integer runTestsCount;
 	private Set<String> failedTests;
 	private BuildStatus buildStatus;
